==> Scripts/08_HDBN_Test.R <==
#############################################################################################
#                                                                                           #
#                                                                                           #
#############           Construct dynamic bayesian network                 #########
#                                                                                           #
#                                                                                           #
#############################################################################################


# This script reads the template HDBN, the CPTs for each core model nodes and the template CPTs for each biological compartments that will be repeated for each variables

# Helpers
# setCPT(net, node, cpt_df) (inverse of getCPT)

# archetype → replica CPT propagation

# CPT structural equivalence checks
